+++ backend/src/main/kotlin/ru/nsu/ctf/paketnikback/domain/service/PcapProcessorService.kt
@@ -1,5 +1,7 @@
 package ru.nsu.ctf.paketnikback.domain.service
 
+import ru.nsu.ctf.paketnikback.domain.dto.RegexSearchMatch
+
 interface PcapProcessorService {
-    // TODO: define methods
+    fun searchByRegex(pcapId: String, regex: Regex): List<RegexSearchMatch>
 }
